requirements.txt
pandas
ragas
datasets
python-dotenv
langchain-openai
langchain-community
PyPDF2
evaluate
bert_score  # Required for BERTScore metric
# Optional: tiktoken (for token counting, improves chunking accuracy)
tiktoken
pytest
faiss-cpu
langchain-core
langchain-text-splitters 